Move GUI logic to its own file. Simplify main fn.
